+++ packages/graphql/src/schema/make-augmented-schema.ts
@@ -75,7 +75,7 @@ import { filterTruthy } from "../utils/utils";
 import { createConnectionFields } from "./create-connection-fields";
 import { addGlobalNodeFields } from "./create-global-nodes";
 import { createRelationshipFields } from "./create-relationship-fields/create-relationship-fields";
-import { deprecationMap } from "./deprecation-map";
+import { bookmarkDeprecationMap } from "./deprecation-map";
 import { AugmentedSchemaGenerator } from "./generation/AugmentedSchemaGenerator";
 import { withAggregateSelectionType } from "./generation/aggregate-types";
 import { withCreateInputType } from "./generation/create-input";
@@ -84,6 +84,7 @@ import { withObjectType } from "./generation/object-type";
 import { withMutationResponseTypes } from "./generation/response-types";
 import { withOptionsInputType } from "./generation/sort-and-options-input";
 import { withUpdateInputType } from "./generation/update-input";
+import { shouldAddDeprecatedFields } from "./generation/utils";
 import { withUniqueWhereInputType, withWhereInputType } from "./generation/where-input";
 import getNodes from "./get-nodes";
 import { getResolveAndSubscriptionMethods } from "./get-resolve-and-subscription-methods";
@@ -151,11 +152,16 @@ function makeAugmentedSchema({
     }
 
     // Loop over all entries in the deprecation map and add field deprecations to all types in the map.
-    for (const [typeName, deprecatedFields] of deprecationMap) {
+    for (const [typeName, deprecatedFields] of bookmarkDeprecationMap) {
         const typeComposer = composer.getOTC(typeName);
-        typeComposer.deprecateFields(
-            deprecatedFields.reduce((acc, { field, reason }) => ({ ...acc, [field]: reason }), {})
-        );
+
+        if (shouldAddDeprecatedFields(features, "bookmark")) {
+            typeComposer.deprecateFields(
+                deprecatedFields.reduce((acc, { field, reason }) => ({ ...acc, [field]: reason }), {})
+            );
+        } else {
+            typeComposer.removeField(deprecatedFields.map((field) => field.field));
+        }
     }
 
     // TODO: ideally move these in getSubgraphSchema()
@@ -248,6 +254,7 @@ function makeAugmentedSchema({
                 propagatedDirectivesForNode,
                 aggregationTypesMapper,
                 seenRelationshipPropertiesTypes,
+                features,
             });
             const connectionFields = createConnectionFields({
                 entityAdapter: interfaceEntityAdapter,
@@ -651,7 +658,7 @@ function generateInterfaceObjectType({
 }: {
     composer: SchemaComposer;
     interfaceEntityAdapter: InterfaceEntityAdapter;
-    features?: Neo4jFeaturesSettings;
+    features: Neo4jFeaturesSettings | undefined;
     subgraph?: Subgraph;
     userDefinedFieldDirectivesForNode: Map<string, Map<string, DirectiveNode[]>>;
     userDefinedInterfaceDirectives: DirectiveNode[];