+++ packages/graphql/tests/tck/types/datetime.test.ts
@@ -110,8 +110,7 @@ describe("Cypher DateTime", () => {
                             \\"timeZoneOffsetSeconds\\": 0
                         }
                     }
-                ],
-                \\"resolvedCallbacks\\": {}
+                ]
             }"
         `);
     });