+++ packages/graphql/tests/tck/types/duration.test.ts
@@ -149,8 +149,7 @@ describe("Cypher Duration", () => {
                             }
                         }
                     }
-                ],
-                \\"resolvedCallbacks\\": {}
+                ]
             }"
         `);
     });